+++ src/deepshogi/native/cpp/ThreadPool.cpp
@@ -1,7 +1,5 @@
 #include "ThreadPool.h"
 
-#include <iostream>
-
 namespace deepshogi {
 
 /**
@@ -75,7 +73,7 @@ void ThreadPool::_run() {
  * @return Number of threads
  */
 int32_t ThreadPool::getSize() {
-  return _threads.size();
+  return (int32_t)_threads.size();
 }
 
 }  // namespace deepshogi
